test(docker): move Docker-dependent tests behind integration build tag
